Minor fixes to flash/fs stuff
- Implemented basic flash init

  This reads the superblock, formats the flash if it is empty, and
  validates the read superblock before doing anything significant with
  it.

- Add flash helpers for resetting the device

- Reliably add newlines for logs on UART if scheduler isn't yet running

- Fixed CRC32 routine; it was previously broken when compiling with
  optimizations and would lead to some ~ funky ~ behavior

- Be a bit less obnoxious about warnings

+++ host-radio/Sources/Util/Crc32.h
@@ -18,7 +18,7 @@ static inline uint32_t Crc32(etl::span<const uint8_t> message) {
     for(size_t i = 0; i < message.size(); i++) {
         byte = message[i];
         crc = crc ^ byte;
-        for(size_t j = 7; j>= 0; j--) {
+        for(int j = 7; j>= 0; j--) {
             mask = -(crc & 1);
             crc = (crc >> 1) ^ (0xEDB88320 & mask);
         }
